Re-organizing code to better separate responsibilities of code.
