+++ chat_bot/__init__.py
@@ -1,0 +1,1 @@
+from chat_bot.Manamoji import Manamoji
